Allow passing in extra state from the custom repository class
This may be needed in case the tables for aggregate root and child entities
don't containa ll the information you may need inside the entity. This might
be considered an issue with the database schema, but you may not be able
to fix the problem. So instead, you need a work-around, which is provided
by this change.